💥 Move stoppers to top level of module
