+++ .flake8
@@ -1,7 +1,7 @@
 [flake8]
 max-line-length = 80
-select = C,E,F,W,B,B950
-ignore = E203, E501, W503
+select = C,E,F,W,B
+ignore = E203, E501, W503, B950
 exclude =
     .git,
     __pycache__,
