+++ docs/source/conf.py
@@ -24,8 +24,16 @@ extensions.append("recommonmark")
 
 autoapi_type = "python"
 autoapi_dirs = ["../../src/rt_stoppers_contrib"]
-autoapi_ignore = ["*/test_*.py"]
+autoapi_ignore = ["*/no_improvement.py", "*/threshold_by_epoch.py"]
 autoapi_python_class_content = "init"
+autoapi_options = [
+    "members",
+    "undoc-members",
+    "show-inheritance",
+    "show-module-summary",
+    "special-members",
+    "imported-members",
+]
 
 templates_path = ["_templates"]
 exclude_patterns = []
